Fx bug where sensor baud rate was saved in a byte instead of an int

+++ SDS021.h
@@ -77,7 +77,7 @@ private:
 	SoftwareSerial SoftwareSerial_;
 	
 	// Message properties
-	static const byte kBaudRate_ = 9600;
+	static const int kBaudRate_ = 9600;
 	static const byte kOutputLength_ = 19;
 	static const byte kInputLength_ = 10;
